// src/app/Models/publicacion.ts
export class Publicacion{
    id_publicacion: number;
    titulo:string;
    fecha_publicacion: Date;
    descripcion:string;
    enlace: string;
    profesion:string;
    estado_profesion: string;
    ruta_archivo:string;
    tipo_archivo: string;
    id_tipo_publicacion: number;
    id_usuario: number;
    id_estado_publicacion: number;
    nombre_perfil:string;
}